Organization to allow easier implementation for components
