+++ src/components/groups/group-list.js
@@ -1,0 +1,21 @@
+import GroupItem from "./group-item";
+
+const GroupsList = (
+    {
+        groups =
+            []
+    }
+) => {
+    return (
+        <>
+            <ul className={'list-group'}>
+                {
+                    groups.map && groups.map(group =>
+                                                 <GroupItem key={group._id} group={group}/>)
+                }
+            </ul>
+        </>
+    )
+}
+
+export default GroupsList
